address compiler warnings

+++ Breakout/GameManager.cpp
@@ -8,7 +8,7 @@ GameManager::GameManager(sf::RenderWindow* window)
     _messagingSystem(nullptr), _ui(nullptr), _pause(false), _time(0.f), _lives(3), _pauseHold(0.f), _levelComplete(false),
     _powerupInEffect({ none,0.f }), _timeLastPowerupSpawned(0.f),
     _shakeTimeRemaining(0.f),_shakeIntensity(0.f),_shakeOffset({0.f, 0.f}),
-    _lastMouseX(0.f)
+    _lastMouseX(0)
 {
     _font.loadFromFile("font/montS.ttf");
     _masterText.setFont(_font);
@@ -19,7 +19,7 @@ GameManager::GameManager(sf::RenderWindow* window)
 
 void GameManager::initialize()
 {
-    srand(time(0));
+    srand(static_cast<uint32_t>(time(0)));
     _paddle = std::make_unique<Paddle>(_window);
     _brickManager = std::make_unique<BrickManager>(_window, this);
     _messagingSystem = std::make_unique<MessagingSystem>(_window);
@@ -95,7 +95,7 @@ void GameManager::update(float dt)
     if (sf::Mouse::getPosition(*_window).x != _lastMouseX)
     {
         sf::Vector2i mousePos = sf::Mouse::getPosition(*_window);
-        _paddle->setPosition(mousePos.x);
+        _paddle->setPosition(static_cast<float>(mousePos.x));
         _lastMouseX = mousePos.x;
     }
     else
